aws/vpc/delete-all-default-vpcs/cdk/requirements.txt
aws-cdk.core
aws_cdk.aws_events
aws_cdk.aws_events_targets
aws_cdk.aws_iam
aws-cdk.aws-lambda
aws_cdk.aws_logs
aws_cdk.aws_ssm
